plots of C19 convalescent efficacy vs. time
added plotting functions for C19 convalescent protection as a function of time as per Khoury et al.. Also explored allowing lambda (neut decay rate) to vary from person to person - support on [0, inf) of lambda produces neut distributions that deviate from lognormal as time goes on, but the extent of this is a function of the variance in lambda. Neut distributions also get wider as time goes on which makes bimodality more robust.